edit palette variations in modal
