+++ src/components/VariationGraph/ScalarVariationGraph.tsx
@@ -2,7 +2,6 @@ import { Box, Text, useToken } from "@chakra-ui/react";
 import { LineChart, Line, Tooltip, YAxis } from "recharts";
 import { Variation } from "@/src/types/Variations/Variation";
 import { Block } from "@/src/types/Block";
-import { memo } from "react";
 import { VARIATION_BOUND_WIDTH } from "@/src/utils/layout";
 import { SplineVariation } from "@/src/types/Variations/SplineVariation";
 import { SplineVariationGraph } from "@/src/components/VariationGraph/SplineVariationGraph";
@@ -16,7 +15,7 @@ type ScalarVariationGraphProps = {
   block: Block;
 };
 
-export const ScalarVariationGraph = memo(function ScalarVariationGraph({
+export const ScalarVariationGraph = function ScalarVariationGraph({
   uniformName,
   variation,
   width,
@@ -64,7 +63,7 @@ export const ScalarVariationGraph = memo(function ScalarVariationGraph({
       </LineChart>
     </Box>
   );
-});
+};
 
 const ScalarValueTooltip = ({
   active,
